Checkpoint started changing views
